PM-110 add: Add logic export file ppt
