--- pm-app/src/data/output/output.json
@@ -1,1 +1,0 @@
-{}
